requirements.txt
line-bot-sdk
bs4
flask
datetime
emoji
pyshorteners
requests
gunicorn
pymongo